Updating app structure.
I've decided it would be best practice for App to house all the state.
(Actions, Nodes, etc) (All the stuff from API requests)

Created new HOC withNewFetching which takes another argument to save the data to.
Passing "actions" as the second argument does a couple things.
Data is saved to actionsData,
isLoading is saved to actionsIsLoading,
and errors are saved to actionsError.

This allows for multiple requests to be done is parallel.

Created new file Action.js and Actions.js for the new Action list
and Action object. These classes have the infomation passed a
props/children instead of fetching it themselves.

(At this point the API has also changed and ActionList.js no longer works)